Refactor crypto-list
