+++ components/crypto-list/types.ts
@@ -1,0 +1,4 @@
+import type { CryptoListingsResponse } from "../../hooks/useCryptoListings"
+
+export type CryptoItem = CryptoListingsResponse["data"][number]
+export type SortKey = "name" | "symbol" | "price"
